+++ stats_test.go
@@ -813,3 +813,29 @@ func TestHarmonicMean(t *testing.T) {
 		t.Errorf("Empty slice should have returned an error")
 	}
 }
+
+func TestCovariance(t *testing.T) {
+	s1 := []float64{1, 2, 3, 4, 5}
+	s2 := []float64{10, -51.2, 8}
+	s3 := []float64{1, 2, 3, 5, 6}
+	s4 := []float64{}
+
+	_, err := Covariance(s1, s2)
+	if err == nil {
+		t.Errorf("Mismatched slice lengths should have returned an error")
+	}
+
+	a, err := Covariance(s1, s3)
+	if err != nil {
+		t.Errorf("Should not have returned an error")
+	}
+
+	if a != 3.2499999999999996 {
+		t.Errorf("Covariance %v != %v", a, 3.2499999999999996)
+	}
+
+	_, err = Covariance(s1, s4)
+	if err == nil {
+		t.Errorf("Empty slice should have returned an error")
+	}
+}